+++ src/components/ReaderHeader.tsx
@@ -1,8 +1,12 @@
 import React from "react";
 
+import { RSPrefs } from "../preferences";
+
 import Locale from "../resources/locales/en.json";
 import readerStateStyles from "./assets/styles/readerStates.module.css";
 import readerHeaderStyles from "./assets/styles/readerHeader.module.css";
+
+import TargetIcon from "./assets/icons/target-icon.svg";
 
 import { Links } from "@readium/shared";
 
@@ -14,11 +18,17 @@ import { SettingsAction } from "./SettingsAction";
 import { FullscreenAction } from "./FullscreenAction";
 import { TocAction } from "./TocAction";
 import { OverflowMenu } from "./OverflowMenu";
+import { OverflowMenuItem } from "./templateComponents/OverflowMenuItem";
+
+import parseTemplate from "json-templates";
 
 export const ReaderHeader = ({ runningHead, toc }: { runningHead: string | undefined, toc: Links }) => {
   const isImmersive = useAppSelector(state => state.reader.isImmersive);
   const isHovering = useAppSelector(state => state.reader.isHovering);
   const dispatch = useAppDispatch();
+
+  const jsonTemplate = parseTemplate(RSPrefs.shortcuts.jumpToPosition);
+  const platformModifier = useAppSelector(state => state.reader.platformModifier);
 
   const setHover = () => {
     dispatch(setHovering(true));
@@ -56,7 +66,14 @@ export const ReaderHeader = ({ runningHead, toc }: { runningHead: string | undef
         <SettingsAction />
         <FullscreenAction />
         <TocAction toc={ toc } />
-        <OverflowMenu />
+        <OverflowMenu>
+          <OverflowMenuItem
+            SVG={ TargetIcon } 
+            label={ Locale.reader.jumpToPosition.label }
+            shortcut={ jsonTemplate({ PlatformKey: platformModifier.icon }) } 
+            onActionCallback={ () => {} }
+          />
+        </OverflowMenu>
       </div>
     </header>
     </>
