+++ src/components/templateComponents/OverflowMenuItem.tsx
@@ -1,13 +1,13 @@
 import React, { ComponentType, SVGProps } from "react";
 
-import overflowMenuStyles from "./assets/styles/overflowMenu.module.css";
+import overflowMenuStyles from "../assets/styles/overflowMenu.module.css";
 
 import { Keyboard, MenuItem, Text } from "react-aria-components";
 
 export interface IOverflowMenuItemProp {
   SVG: ComponentType<SVGProps<SVGElement>>;
   label: string;
-  shortcut: string;
+  shortcut?: string;
   onActionCallback: () => void;
 }
 
@@ -22,7 +22,7 @@ export const OverflowMenuItem = ({
       <MenuItem className={ overflowMenuStyles.menuItem } onAction={ onActionCallback }>
         <SVG aria-hidden="true" focusable="false" />
         <Text className={ overflowMenuStyles.menuItemLabel } slot="label">{ label }</Text>
-        <Keyboard className={ overflowMenuStyles.menuItemKbdShortcut }>{ shortcut }</Keyboard>
+      { shortcut && <Keyboard className={ overflowMenuStyles.menuItemKbdShortcut }>{ shortcut }</Keyboard> }
       </MenuItem>
     </>
   )
